rename: Link to LinkItem
